Add ChannelMetadata and ChannelMetadataCollection classes; implement ChannelProtocol interface
